+++ src/gView.Framework.Core/Symbology/ILegendGroup.cs
@@ -1,4 +1,6 @@
-﻿namespace gView.Framework.Core.Symbology
+﻿using gView.Framework.Core.Data;
+
+namespace gView.Framework.Core.Symbology
 {
     public interface ILegendGroup
     {
@@ -6,4 +8,13 @@
         ILegendItem LegendItem(int index);
         void SetSymbol(ILegendItem item, ISymbol symbol);
     }
+
+    public interface ILegendDependentFields
+    {
+        string[] LegendDependentFields { get; }
+        string LegendSymbolOrderField { get; }
+
+        string LegendSymbolKeyFromFeature(IFeature feature);
+        ILegendItem LegendItemFromSymbolKey(string symbolKey);
+    } 
 }