+++ src/gView.Interoperability.GeoServices/Rest/DTOs/Request/JsonQueryLegendsDTO.cs
@@ -1,0 +1,5 @@
+﻿namespace gView.Interoperability.GeoServices.Rest.DTOs.Request
+{
+    public class JsonQueryLegendsDTO : JsonExportMapDTO 
+    { }
+}